Refactor HTTP serialization (API Change):
A new class is introduced to permit incremental serialization
of HTTP messages. Existing free functions are re-implemented in
terms of this new implementation.

The Body::writer concept is refined to support a greater variety
of strategies for providing buffers representing the body to
the serialization algorithms.

Added buffer_body, a new model of Body which allows the caller
to provide a series of owned bufers using their own serialization
loop.

Added empty_body, a model of Body which is for serialization only,
to represent HTTP messages with no content body.